Prototype finishes
Rewrote the UI system to handle arbitrary numbers of attacks on a per player basis. Target selection is done through a drop down and an outline shader is applied to the chosen target.  Added an info screen for displaying details of a selected attack and enemy data.(still need one for players) A win or lose state is detected by the battle manager. In the future this will move on to another screen right now it kills the program. Balanced the stats of all entities. Added an effects system. Effects can be applied to an attack along with a proc chance. When an effect procs it is added to the target and will do its thing. Added several new effects. Added several new attacks using the new effects. All entities are currently given two random attacks (duplicates allowed).